embeddings/pyproject.toml
[project]
name = "embeddings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary>=2.9.12",
    "requests>=2.33.1",
]
